+++ src/parser/mod.rs
@@ -11,6 +11,6 @@ macro_rules! parser_test {
 
 mod tokens;
 mod match_;
-mod decl;
 
+pub mod decl;
 pub mod expr;